chore: change /messagebox folder structure
